requirements.txt
astroid==2.0.4
certifi==2019.6.16
chardet==3.0.4
dataclasses==0.6
dj-database-url==0.5.0
Django==2.1.2
django-heroku==0.3.1
djangorestframework==3.8.2
djongo==1.2.30
dnspython==1.16.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.15.2
pandas==0.23.4
Pillow==6.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pylint==2.1.1
pymongo==3.7.2
PyPDF2==1.26.0
python-dateutil==2.7.3
pytz==2018.5
reportlab==3.5.21
requests==2.22.0
six==1.11.0
sqlparse==0.2.4
stripe==2.32.1
urllib3==1.25.3
webencodings==0.5.1
whitenoise==4.1.2
wrapt==1.10.11
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
